IDP_compressed/Unbordered/Xerox_Table/requirements.txt
werkzeug==2.2.1
FLASK==2.2.1
pymongo==4.2.0
timm==0.6.5
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
opencv-python==4.5.3.56
pdf2image==1.16.0
scipy==1.7.3
flask_cors==3.0.10
pandas==1.3.3
pyenchant==3.2.2
pillow==9.0.1
numpy==1.22.0
tesserocr==2.5.2
matplotlib==3.4.3
efficientnet==1.1.1
imgaug==0.4.0
extract_msg==0.28.7
textract==1.6.5
boto3==1.17.109
tensorflow==2.7.2
validators==0.20.0
setuptools==65.0.2
spacy==3.4.1
transformers==4.21.1
scikit-learn==1.1.2
fasttext==0.9.2
redis==4.3.4
regex==2022.8.17
tensorboard==2.10.0
zipp==3.8.1
celery==5.2.7
pytorch-lightning==1.7.5
lmdb==1.3.0
nltk==3.7
gensim==3.8.3
kneed==0.8.1
docx2pdf==0.1.8
pdfkit==1.0.0
spacy-alignments==0.8.5
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
gunicorn==20.1.0
imutils==0.5.4